tests: port helperSpec to jest
our first jest based tests :)

PR: #815
PR-URL: https://github.com/apache/couchdb-fauxton/pull/815
